Agregando la capa framework (close #23)
