+++ app/src/main/java/com/architect/coders/pokedex/data/PhotoRepository.kt
@@ -1,0 +1,13 @@
+package com.architect.coders.pokedex.data
+
+import android.net.Uri
+import arrow.core.Either
+
+interface PhotoRepository {
+
+    val path: String
+
+    fun createFile(nameFile: String): Either<Error, Uri>
+
+    fun deleteImageFile(fileName: String): Error?
+}
