--- app/src/main/java/com/architect/coders/pokedex/data/database/TypeL.kt
@@ -1,11 +1,0 @@
-package com.architect.coders.pokedex.data.database
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "type")
-data class TypeL(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val pokemonID: Int,
-    val name: String
-)
