+++ app/src/main/java/com/architect/coders/pokedex/framework/database/PokemonDAO.kt
@@ -1,4 +1,4 @@
-package com.architect.coders.pokedex.data.database
+package com.architect.coders.pokedex.framework.database
 
 import androidx.room.*
 import kotlinx.coroutines.flow.Flow
