+++ app/src/main/java/com/architect/coders/pokedex/framework/database/StatL.kt
@@ -1,0 +1,12 @@
+package com.architect.coders.pokedex.framework.database
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity(tableName = "stat")
+data class StatL(
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    val pokemonID: Int,
+    val name: String,
+    val baseStat: Int
+)
