+++ app/src/main/java/com/architect/coders/pokedex/framework/network/TypeR.kt
@@ -1,4 +1,4 @@
-package com.architect.coders.pokedex.data.network
+package com.architect.coders.pokedex.framework.network
 
 import com.google.gson.annotations.SerializedName
