+++ app/src/main/java/com/architect/coders/pokedex/framework/network/TypeX.kt
@@ -1,8 +1,8 @@
-package com.architect.coders.pokedex.data.network
+package com.architect.coders.pokedex.framework.network
 
 import com.google.gson.annotations.SerializedName
 
-data class PokemonItemR(
+data class TypeX(
     @SerializedName("name")
     val name: String,
     @SerializedName("url")
