+++ app/src/main/java/com/architect/coders/pokedex/usecases/GetPathUseCase.kt
@@ -1,7 +1,7 @@
 package com.architect.coders.pokedex.usecases
 
-import com.architect.coders.pokedex.data.FileRepository
+import com.architect.coders.pokedex.data.PhotoRepository
 
-class GetPathUseCase(private val fileRepository: FileRepository) {
-    operator fun invoke(): String = fileRepository.path
+class GetPathUseCase(private val photoRepository: PhotoRepository) {
+    operator fun invoke(): String = photoRepository.path
 }
